fix: APP-732 name and folder fixes, moved functions to util file
