+++ web-marketplace/src/components/organisms/OrganizationMembers/OrganizationMembers.constants.tsx
@@ -40,6 +40,7 @@ export const REGEN_ADDRESS_REQUIRED_ERROR =
   'You must enter a REGEN address in order to add this user as an Admin or Editor.';
 export const INVALID_EMAIL_ERROR = 'Invalid email';
 export const INVALID_REGEN_ADDRESS_ERROR = 'Invalid REGEN address';
+
 // Remove Member Modal
 export const REMOVE_MEMBER_TITLE = msg`Are you sure you want to remove this member from your organization?`;
 export const REMOVE_MEMBER_DESCRIPTION = msg`Any posts they have created or signed will remain.`;
