+++ web-marketplace/src/components/organisms/OrganizationMembers/OrganizationMembers.types.tsx
@@ -1,3 +1,5 @@
+import { GetAccountsByNameOrAddrQuery } from 'generated/graphql';
+
 import { BaseMemberRole } from '../BaseMembersTable/BaseMembersTable.types';
 
 export type Member = {
@@ -13,6 +15,18 @@ export type Member = {
   hasWalletAddress: boolean;
   invited?: boolean;
 };
+
+export interface InviteMemberModalProps {
+  open: boolean;
+  onClose: () => void;
+  onSubmit: (data: {
+    role: BaseMemberRole | undefined;
+    addressOrEmail: string;
+    visible: boolean;
+  }) => void;
+  accounts?: GetAccountsByNameOrAddrQuery | null;
+  setDebouncedValue?: (value: string) => void;
+}
 
 export type VisibilitySwitchProps = {
   checked: boolean;
