+++ web-marketplace/src/components/organisms/OrganizationMembers/OrganizationMembers.utils.tsx
@@ -1,9 +1,9 @@
-import { MessageDescriptor } from '@lingui/core';
-
 import { CogIcon } from 'web-components/src/components/icons/CogIcon';
 import EditIcon from 'web-components/src/components/icons/EditIcon';
 import EyeIcon from 'web-components/src/components/icons/EyeIcon';
 import { OwnerIcon } from 'web-components/src/components/icons/OwnerIcon';
+
+import { TranslatorType } from 'lib/i18n/i18n.types';
 
 import {
   ROLE_ADMIN,
@@ -11,10 +11,7 @@ import {
   ROLE_OWNER,
   ROLE_VIEWER,
 } from '../ActionDropdown/ActionDropdown.constants';
-import {
-  BaseMemberRole,
-  RoleOption,
-} from '../BaseMembersTable/BaseMembersTable.types';
+import { RoleOption } from '../BaseMembersTable/BaseMembersTable.types';
 import {
   ROLE_ADMIN_DESCRIPTION,
   ROLE_ADMIN_LABEL,
@@ -25,7 +22,6 @@ import {
   ROLE_VIEWER_DESCRIPTION,
   ROLE_VIEWER_LABEL,
 } from './OrganizationMembers.constants';
-import { TranslatorType } from 'lib/i18n/i18n.types';
 
 export const getRoleItems = (_: TranslatorType) =>
   [
